requirements.txt
psutil
supabase
requests
ping3
WinTmp
GPUtil
speedtest-cli